Split the transition file in dedicated components
