+++ src/components/Translate.js
@@ -5,47 +5,6 @@ import { Transition } from 'react-transition-group';
 const DURATION = 375;
 const BASE_TRANSITION = `${DURATION}ms ease-in-out`;
 
-const fadeDefaultStyle = {
-    opacity: 0,
-    visibility: 'hidden',
-    transition: `opacity ${BASE_TRANSITION}`,
-};
-
-const fadeTransitionStyles = {
-    entering: { opacity: 0, visibility: 'hidden' },
-    entered: { opacity: 1, visibility: 'visible' },
-    exiting: { opacity: 1, visibility: 'visible' },
-    exited: { opacity: 0, visibility: 'hidden' },
-};
-
-export class Fade extends PureComponent {
-    static propTypes = {
-        children: PropTypes.oneOfType([PropTypes.element, PropTypes.node])
-            .isRequired,
-        timeout: PropTypes.number,
-    };
-
-    static defaultProps = {
-        timeout: DURATION,
-    };
-
-    render() {
-        const { children, ...props } = this.props;
-        return (
-            <Transition {...props}>
-                {state =>
-                    React.cloneElement(children, {
-                        style: {
-                            ...fadeDefaultStyle,
-                            ...fadeTransitionStyles[state],
-                        },
-                    })
-                }
-            </Transition>
-        );
-    }
-}
-
 const translateDefaultStyle = {
     transform: 'translate(0, 0)',
     transition: `transform ${BASE_TRANSITION}`,
@@ -56,7 +15,7 @@ const translateTransitionStyles = (x = 0, y = 0) => ({
     entered: { transform: `translate(${x}, ${y})` },
 });
 
-export class Translate extends PureComponent {
+export default class Translate extends PureComponent {
     static propTypes = {
         children: PropTypes.oneOfType([PropTypes.element, PropTypes.node])
             .isRequired,
